Refactor CustomAuthenticationHandler and RequestContext for improved authentication
