+++ server-spi/src/main/java/com/flipkart/varadhi/spi/RequestContext.java
@@ -1,14 +1,16 @@
-package com.flipkart.varadhi.entities.utils;
+package com.flipkart.varadhi.spi;
+
 
 import lombok.Data;
 
 import java.net.URI;
 import java.util.Map;
+import io.vertx.core.MultiMap;
 
 @Data
 public class RequestContext {
     private URI uri;
-    private Map<String, String> params;
-    private Map<String, String> headers;
+    private MultiMap params;
+    private MultiMap headers;
     private Map<String, Object> context;
 }
