+++ server-spi/src/main/java/com/flipkart/varadhi/spi/authn/Authenticator.java
@@ -1,12 +1,12 @@
 package com.flipkart.varadhi.spi.authn;
 
-import com.flipkart.varadhi.entities.Org;
+
 import com.flipkart.varadhi.entities.auth.UserContext;
-import com.flipkart.varadhi.entities.utils.RequestContext;
+import com.flipkart.varadhi.spi.RequestContext;
 import io.vertx.core.Future;
 
 public interface Authenticator {
     Future<Boolean> init(AuthenticationOptions authenticationOptions);
 
-    Future<UserContext> authenticate(Org org, RequestContext requestContext);
+    Future<UserContext> authenticate(String orgName, RequestContext requestContext);
 }
